Added beforeSwipe callback
Similar to 'callback' this function is called before the actual slide
transition takes place. Rationale: manipulate slide content just before
the slide is displayed.

Note: this doesn't work very smooth if there is a considerable amount
of content manipulation or external media needs to be loaded.

Note2: the "event" parameter is always null.